# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Avviare Educational Hub API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export type AdminUserRole = (typeof AdminUserRole)[keyof typeof AdminUserRole];

export const AdminUserRole = {
  admin: "admin",
  admissions_officer: "admissions_officer",
  content_manager: "content_manager",
} as const;

export interface AdminUser {
  id: number;
  username: string;
  role: AdminUserRole;
  name: string;
}

export interface LoginResponse {
  token: string;
  user: AdminUser;
}

export interface MessageResponse {
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export type ApplicationClassType =
  (typeof ApplicationClassType)[keyof typeof ApplicationClassType];

export const ApplicationClassType = {
  regular: "regular",
  weekend: "weekend",
} as const;

export type ApplicationStatus =
  (typeof ApplicationStatus)[keyof typeof ApplicationStatus];

export const ApplicationStatus = {
  pending: "pending",
  reviewing: "reviewing",
  accepted: "accepted",
  rejected: "rejected",
} as const;

export interface Application {
  id: number;
  name: string;
  email: string;
  phone: string;
  course: string;
  classType: ApplicationClassType;
  message?: string | null;
  status: ApplicationStatus;
  createdAt: string;
  updatedAt: string;
}

export type CreateApplicationBodyClassType =
  (typeof CreateApplicationBodyClassType)[keyof typeof CreateApplicationBodyClassType];

export const CreateApplicationBodyClassType = {
  regular: "regular",
  weekend: "weekend",
} as const;

export interface CreateApplicationBody {
  name: string;
  email: string;
  phone: string;
  course: string;
  classType: CreateApplicationBodyClassType;
  message?: string;
}

export type UpdateApplicationStatusBodyStatus =
  (typeof UpdateApplicationStatusBodyStatus)[keyof typeof UpdateApplicationStatusBodyStatus];

export const UpdateApplicationStatusBodyStatus = {
  pending: "pending",
  reviewing: "reviewing",
  accepted: "accepted",
  rejected: "rejected",
} as const;

export interface UpdateApplicationStatusBody {
  status: UpdateApplicationStatusBodyStatus;
  notes?: string;
}

export interface ApplicationListResponse {
  data: Application[];
  total: number;
}

export type ContactStatus = (typeof ContactStatus)[keyof typeof ContactStatus];

export const ContactStatus = {
  new: "new",
  read: "read",
  replied: "replied",
} as const;

export interface Contact {
  id: number;
  name: string;
  email: string;
  phone: string;
  message: string;
  course?: string | null;
  classType?: string | null;
  status: ContactStatus;
  createdAt: string;
  updatedAt: string;
}

export type CreateContactBodyClassType =
  (typeof CreateContactBodyClassType)[keyof typeof CreateContactBodyClassType];

export const CreateContactBodyClassType = {
  regular: "regular",
  weekend: "weekend",
} as const;

export interface CreateContactBody {
  name: string;
  email: string;
  phone: string;
  message: string;
  course?: string;
  classType?: CreateContactBodyClassType;
}

export type UpdateContactStatusBodyStatus =
  (typeof UpdateContactStatusBodyStatus)[keyof typeof UpdateContactStatusBodyStatus];

export const UpdateContactStatusBodyStatus = {
  new: "new",
  read: "read",
  replied: "replied",
} as const;

export interface UpdateContactStatusBody {
  status: UpdateContactStatusBodyStatus;
}

export interface ContactListResponse {
  data: Contact[];
  total: number;
}

export type CareerStatus = (typeof CareerStatus)[keyof typeof CareerStatus];

export const CareerStatus = {
  pending: "pending",
  reviewing: "reviewing",
  hired: "hired",
  rejected: "rejected",
} as const;

export interface Career {
  id: number;
  name: string;
  email: string;
  phone: string;
  position?: string | null;
  cvUrl?: string | null;
  status: CareerStatus;
  createdAt: string;
  updatedAt: string;
}

export interface CreateCareerBody {
  name: string;
  email: string;
  phone: string;
  position?: string;
  cvUrl?: string;
}

export type UpdateCareerStatusBodyStatus =
  (typeof UpdateCareerStatusBodyStatus)[keyof typeof UpdateCareerStatusBodyStatus];

export const UpdateCareerStatusBodyStatus = {
  pending: "pending",
  reviewing: "reviewing",
  hired: "hired",
  rejected: "rejected",
} as const;

export interface UpdateCareerStatusBody {
  status: UpdateCareerStatusBodyStatus;
}

export interface CareerListResponse {
  data: Career[];
  total: number;
}

export type DashboardStatsApplicationsByStatus = { [key: string]: number };

export type DashboardStatsApplicationsByCourseItem = {
  course: string;
  count: number;
};

export interface DashboardStats {
  totalApplications: number;
  pendingApplications: number;
  totalContacts: number;
  newContacts: number;
  totalCareers: number;
  applicationsByStatus: DashboardStatsApplicationsByStatus;
  applicationsByCourse: DashboardStatsApplicationsByCourseItem[];
}

export type RecentActivityItemsItemType =
  (typeof RecentActivityItemsItemType)[keyof typeof RecentActivityItemsItemType];

export const RecentActivityItemsItemType = {
  application: "application",
  contact: "contact",
  career: "career",
} as const;

export type RecentActivityItemsItem = {
  type: RecentActivityItemsItemType;
  id: number;
  name: string;
  action: string;
  createdAt: string;
};

export interface RecentActivity {
  items: RecentActivityItemsItem[];
}

export type ListApplicationsParams = {
  status?: ListApplicationsStatus;
  course?: string;
  limit?: number;
  offset?: number;
};

export type ListApplicationsStatus =
  (typeof ListApplicationsStatus)[keyof typeof ListApplicationsStatus];

export const ListApplicationsStatus = {
  pending: "pending",
  reviewing: "reviewing",
  accepted: "accepted",
  rejected: "rejected",
} as const;

export type ListContactsParams = {
  status?: ListContactsStatus;
  limit?: number;
  offset?: number;
};

export type ListContactsStatus =
  (typeof ListContactsStatus)[keyof typeof ListContactsStatus];

export const ListContactsStatus = {
  new: "new",
  read: "read",
  replied: "replied",
} as const;

export type ListCareersParams = {
  status?: ListCareersStatus;
  limit?: number;
  offset?: number;
};

export type ListCareersStatus =
  (typeof ListCareersStatus)[keyof typeof ListCareersStatus];

export const ListCareersStatus = {
  pending: "pending",
  reviewing: "reviewing",
  hired: "hired",
  rejected: "rejected",
} as const;