# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Avviare Educational Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current admin user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List all contact inquiries
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [new, read, replied]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactListResponse"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Submit a contact inquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactBody"
      responses:
        "201":
          description: Contact submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/{id}:
    patch:
      operationId: updateContactStatus
      tags: [contacts]
      summary: Update contact inquiry status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContactStatusBody"
      responses:
        "200":
          description: Updated contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
...
components:
  schemas:
...
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/AdminUser"
      required: [token, user]
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [admin, admissions_officer, content_manager]
        name:
          type: string
      required: [id, username, role, name]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Contact:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        message:
          type: string
        course:
          type: string
          nullable: true
        classType:
          type: string
          nullable: true
        status:
          type: string
          enum: [new, read, replied]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, email, phone, message, status, createdAt, updatedAt]
    CreateContactBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        message:
          type: string
        course:
          type: string
        classType:
          type: string
          enum: [regular, weekend]
      required: [name, email, phone, message]
    UpdateContactStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [new, read, replied]
      required: [status]
    ContactListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Contact"
        total:
          type: integer
      required: [data, total]
...